make-dist: revert previous change
That is, do not modify the package name as extracted from
the PackageInfo.g. If a user wants the archive name to
use a different case (e.g. for the "IO" package we want
to use "io" for the dist tarballs), then they can (and should)
simply pass an explicit package name to make-dist.