Check for presence of gh-pages clone first; do not try to create it
It's much more convenient to report such errors first, before the user
had to wait for all kind of other work to complete.

Also, trying to clone the directory is problematic -- e.g. what if the
user is not using SSH, or is using a custom location of the page?