Do not specify remote when doing "git push"
Instead, we rely on a the clone to be properly configured, and an
upstream set for the branch.

This ensures things will work even for people who do use https instead
of git+SSH for push.

If this ever causes problems, we could add a `--git-remote VALUE`
option or so.