Remove -t / --token and -r / --repository
They were broken (in that they had no effect, as the data from
PackageInfo.g either overrode them unconditionally (for the repository)
or was enforced to be a 100% match (for the tag).

Since nobody complained, I am assuming nobody used them. Since I can't
think of a good usecase anymore (now that extract all this data from
the ArchiveURL), removing them seems the best way to go forward. If at
some point somebody wants to have them back, I'll be happy to consider
that as well, based on discussions with the people requesting it.