Allow to specify a custom release script
to support monorepos with only a single release script.

Background: TMP_DIR is created from SRC_DIR via `git archive`, so the
default `.release` script must be checked into git. Thus, in the case of a
monorepo with a single release script it would not be sufficient to copy
the release script to each SRC_DIR before calling `release-gap-package`.
Hence this new option.